+++ src/utils.h
@@ -1,2 +1,2 @@
 #pragma once
-double *gen_kernel(int kernel_size, double sigma);
+float *gen_kernel(int kernel_size, double sigma);
